fixed migrations
