--- backend/database/database/alembic/versions/708db6f21113_initial_migration.py
@@ -1,109 +1,0 @@
-"""Initial migration
-
-Revision ID: 708db6f21113
-Revises: 
-Create Date: 2025-04-21 23:43:09.064805
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '708db6f21113'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('projects',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('title', sa.String(), nullable=False),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('sub_title', sa.String(), nullable=True),
-    sa.Column('abstract', sa.String(), nullable=True),
-    sa.Column('keywords', sa.ARRAY(sa.String()), nullable=True),
-    sa.Column('study_organization', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('users',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('username', sa.String(), nullable=False),
-    sa.Column('first_name', sa.String(), nullable=True),
-    sa.Column('last_name', sa.String(), nullable=True),
-    sa.Column('email', sa.String(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('email'),
-    sa.UniqueConstraint('username')
-    )
-    op.create_table('documents',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('project_id', sa.Integer(), nullable=False),
-    sa.Column('content', sa.Text(), nullable=True),
-    sa.Column('last_modified', sa.DateTime(timezone=True), nullable=False),
-    sa.ForeignKeyConstraint(['project_id'], ['projects.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('project_chapters',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('title', sa.String(), nullable=False),
-    sa.Column('project_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['project_id'], ['projects.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('project_goals',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('content', sa.String(), nullable=False),
-    sa.Column('project_id', sa.Integer(), nullable=False),
-    sa.Column('is_reached', sa.Boolean(), nullable=False),
-    sa.ForeignKeyConstraint(['project_id'], ['projects.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('source_links',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('title', sa.String(), nullable=False),
-    sa.Column('href', sa.String(), nullable=False),
-    sa.Column('project_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['project_id'], ['projects.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('user_to_project',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('project_id', sa.Integer(), nullable=False),
-    sa.Column('role', sa.String(), nullable=False),
-    sa.Column('permissions', sa.String(), nullable=True),
-    sa.Column('last_modified', sa.DateTime(timezone=True), nullable=True),
-    sa.Column('last_viewed', sa.DateTime(timezone=True), nullable=False),
-    sa.ForeignKeyConstraint(['project_id'], ['projects.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('project_sections',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('title', sa.String(), nullable=False),
-    sa.Column('content', sa.Text(), nullable=False),
-    sa.Column('chapter_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['chapter_id'], ['project_chapters.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('project_sections')
-    op.drop_table('user_to_project')
-    op.drop_table('source_links')
-    op.drop_table('project_goals')
-    op.drop_table('project_chapters')
-    op.drop_table('documents')
-    op.drop_table('users')
-    op.drop_table('projects')
-    # ### end Alembic commands ###
