+++ backend/database/database/alembic/versions/89b06bdd11fd_init_tables.py
@@ -1,0 +1,70 @@
+"""init tables
+
+Revision ID: 89b06bdd11fd
+Revises: 
+Create Date: 2025-04-24 17:41:44.488087
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '89b06bdd11fd'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('users',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('username', sa.String(), nullable=False),
+    sa.Column('first_name', sa.String(), nullable=True),
+    sa.Column('last_name', sa.String(), nullable=True),
+    sa.Column('email', sa.String(), nullable=False),
+    sa.Column('password', sa.String(), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('email'),
+    sa.UniqueConstraint('username')
+    )
+    op.create_table('documents',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('user_id', sa.Integer(), nullable=False),
+    sa.Column('title', sa.String(), nullable=False),
+    sa.Column('content', sa.Text(), nullable=False),
+    sa.Column('was_created', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('validations',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('document_id', sa.Integer(), nullable=False),
+    sa.Column('validated', sa.Boolean(), nullable=False),
+    sa.Column('is_valid', sa.Boolean(), nullable=False),
+    sa.ForeignKeyConstraint(['document_id'], ['documents.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('errors',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('validation_id', sa.Integer(), nullable=False),
+    sa.Column('error', sa.String(), nullable=False),
+    sa.Column('resolved', sa.Boolean(), nullable=False),
+    sa.ForeignKeyConstraint(['validation_id'], ['validations.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('errors')
+    op.drop_table('validations')
+    op.drop_table('documents')
+    op.drop_table('users')
+    # ### end Alembic commands ###
