--- backend/database/database/alembic/versions/9cfb078c7d18_deleting_unused_tables.py
@@ -1,51 +1,0 @@
-"""deleting unused tables
-
-Revision ID: 9cfb078c7d18
-Revises: 56bf3faaeaf5
-Create Date: 2025-04-24 17:04:07.647556
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = '9cfb078c7d18'
-down_revision: Union[str, None] = '56bf3faaeaf5'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('project_chapters')
-    op.drop_table('projects')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('projects',
-    sa.Column('id', sa.INTEGER(), server_default=sa.text("nextval('projects_id_seq'::regclass)"), autoincrement=True, nullable=False),
-    sa.Column('title', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('description', sa.TEXT(), autoincrement=False, nullable=True),
-    sa.Column('sub_title', sa.VARCHAR(), autoincrement=False, nullable=True),
-    sa.Column('abstract', sa.VARCHAR(), autoincrement=False, nullable=True),
-    sa.Column('keywords', postgresql.ARRAY(sa.VARCHAR()), autoincrement=False, nullable=True),
-    sa.Column('study_organization', sa.VARCHAR(), autoincrement=False, nullable=True),
-    sa.PrimaryKeyConstraint('id', name='projects_pkey'),
-    postgresql_ignore_search_path=False
-    )
-    op.create_table('project_chapters',
-    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-    sa.Column('title', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('project_id', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.ForeignKeyConstraint(['project_id'], ['projects.id'], name='project_chapters_project_id_fkey'),
-    sa.PrimaryKeyConstraint('id', name='project_chapters_pkey')
-    )
-
-    # ### end Alembic commands ###
