--- backend/database/database/alembic/versions/a65c795405b5_add_password_column_to_users.py
@@ -1,33 +1,0 @@
-"""Add password column to users
-
-Revision ID: a65c795405b5
-Revises: 708db6f21113
-Create Date: 2025-04-22 19:50:17.268227
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = 'a65c795405b5'
-down_revision: Union[str, None] = '708db6f21113'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('password', sa.String(), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('users', 'password')
-
-    # ### end Alembic commands ###
