+++ backend/database/database/database/base.py
@@ -1,6 +1,4 @@
 from ..models.users import User
 from ..models.documents import Document
-from ..models.document_goals import DocumentGoal
-from ..models.document_chapters import DocumentChapter
-from ..models.document_sections import DocumentSection
-from ..models.source_links import SourceLink
+from ..models.validations import Validation
+from ..models.errors import Error
